Create Find-asn-Sections.ps1
Found how to spot the different sections in an .ASN file. Usually there are two identical bytes > 250 (FA FA) followed by two other bytes. The two other bytes are the length of that section. Not sure how to find the first section in any file,  the initial offset of 0x202 works on my 32 team league. 

More testing needs to be done, but this should provide a framework for editing the ASN file.

To do
- find where that first section starts (likely after the big FF FF FF FF ... section)
- figure out what each section does